# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EdeQR LMS API",
        "description": "bridge digital LMS projesi için REST API dokümantasyonu",
        "contact": {
            "email": "ahmet@aciksari.net"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/schools/{school}/manager/days": {
            "get": {
                "tags": [
                    "School Day Management"
                ],
                "summary": "Okulun haftalık açık günlerini listeler",
                "operationId": "listSchoolDays",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "description": "Okul ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "week",
                        "in": "path",
                        "description": "SchoolWeek ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SchoolDay"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Yetkisiz Erişim"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "School Day Management"
                ],
                "summary": "Yeni bir okul günü kaydı oluşturur (Örn: Ek bir gün)",
                "operationId": "storeSchoolDay",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "description": "Okul ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSchoolDayRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SchoolDay"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Yetkisiz Erişim"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/schools/{school}/manager/days/{day}": {
            "get": {
                "tags": [
                    "School Day Management"
                ],
                "summary": "Belirli bir okul gününün detayını getirir",
                "operationId": "showSchoolDay",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "description": "Okul ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "day",
                        "in": "path",
                        "description": "SchoolDay ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SchoolDay"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kayıt Bulunamadı veya Okula Ait Değil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "School Day Management"
                ],
                "summary": "Belirli bir günü günceller (Örn: Açık/Kapalı Ayarı)",
                "operationId": "updateSchoolDay",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "description": "Okul ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "day",
                        "in": "path",
                        "description": "SchoolDay ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSchoolDayRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SchoolDay"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kayıt Bulunamadı"
                    },
                    "403": {
                        "description": "Yetkisiz Erişim"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "School Day Management"
                ],
                "summary": "Belirli bir okul günü kaydını siler",
                "operationId": "deleteSchoolDay",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "description": "Okul ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "day",
                        "in": "path",
                        "description": "SchoolDay ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Başarılı (İçerik Yok)"
                    },
                    "404": {
                        "description": "Kayıt Bulunamadı veya Okula Ait Değil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreSchoolDayRequest": {
                "title": "Okul Günü Oluşturma İsteği",
                "required": [
                    "day_of_week_no",
                    "name"
                ],
                "properties": {
                    "day_of_week_no": {
                        "description": "Haftanın sayısal günü (1=Pazartesi, 7=Pazar)",
                        "type": "integer",
                        "example": 6
                    },
                    "name": {
                        "description": "Günün adı",
                        "type": "string",
                        "example": "Cumartesi"
                    },
                    "is_open": {
                        "description": "Bu günün açık olup olmadığı",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "UpdateSchoolDayRequest": {
                "title": "Okul Günü Güncelleme İsteği",
                "required": [
                    "is_open"
                ],
                "properties": {
                    "is_open": {
                        "description": "Bu günün açık olup olmadığı",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "SchoolDay": {
                "title": "School Day Model",
                "description": "Okulun Haftalık Çalışma Günlerini Temsil Eder.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "school_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "day_of_week_no": {
                        "description": "Haftanın sayısal günü (1=Pazartesi, 7=Pazar)",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Günün adı",
                        "type": "string",
                        "example": "Pazartesi"
                    },
                    "is_open": {
                        "description": "Bu gün okulun açık olup olmadığı",
                        "type": "boolean",
                        "example": true
                    },
                    "start_time": {
                        "description": "Günlük başlangıç saati",
                        "type": "string",
                        "format": "time",
                        "example": "08:30:00",
                        "nullable": true
                    },
                    "end_time": {
                        "description": "Günlük bitiş saati",
                        "type": "string",
                        "format": "time",
                        "example": "17:30:00",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT veya Sanctum tokenini \"Bearer {token}\" formatında girin"
            },
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "1": {
            "bearerAuth": []
        }
    }
}